requirements.txt
Flask
flask-cors
python-dotenv
sqlalchemy
cx_Oracle
flasgger
marshmallow
paramiko
py7zr
pandas